Store order money as integer minor units
Move money through the order domain, logic, saga and database from
float64 dollars to int64 minor units (cents). Integer minor units give
exact arithmetic with no float rounding drift and are the unit the
payment authorize/refund path speaks, so the saga can hand amounts over
without a lossy conversion later.

The HTTP JSON contract stays in dollars: a response DTO renders minor
units back to a dollars float at the serialization boundary, so the
frontend is unaffected. Conversion happens only at the edges — cart
ingress (dollars to cents) and HTTP/notification egress (cents to
dollars) via the domain MinorUnits/Dollars helpers.

Migration 000006 converts the DECIMAL(10,2) columns to BIGINT (scaling
existing rows by 100) and drops/recreates the total and item-subtotal
CHECK constraints, which hold identically in minor units. The demo seed
is converted to minor units to match. Order-create no longer binds
client-supplied items (they are server-sourced from the cart), so a body
carrying dollar prices is ignored rather than rejected by int64 binding.

+++ internal/saga/workflow.go
@@ -34,7 +34,7 @@ type ReserveItem struct {
 type NotifyInput struct {
 	OrderID string
 	UserID  string
-	Total   float64
+	Total   int64 // minor units
 }
 
 // OrderFulfillmentInput is the workflow input. The best-effort cart-clear step
@@ -43,7 +43,7 @@ type NotifyInput struct {
 type OrderFulfillmentInput struct {
 	OrderID string
 	UserID  string
-	Total   float64
+	Total   int64 // minor units
 	Items   []ReserveItem
 }
